requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
Werkzeug==3.0.1
Flask-Migrate==4.0.5
pymupdf>=1.26.0
python-docx>=1.2.0
openai>=2.8.0
fire>=0.7.0
loguru>=0.7.3
datasets>=4.4.0
typer>=0.20.0
rich>=13.7.0
python-dotenv>=1.0.1
requests>=2.31.0
